refactor(proposals): extract analytics/dashboard/alerts into proposal_analytics_service
Move ~1,450 lines of inline aggregation out of views/proposal.py:
build_proposal_analytics, build_dashboard and build_proposal_alerts now
live in content/services/proposal_analytics_service.py together with the
engagement/heat-score helpers and their constants. The three endpoints
become thin permission+Response wrappers, and the view keeps
underscore-aliases for every moved symbol so internal call sites
(track_proposal_engagement, list_proposals, CSV export, dismiss) and test
imports keep working unchanged.

One output-identical query optimization rides along: the per-section-type
latest-title loop in the analytics builder is now a single values_list
pass. tasks.py imports the heat-score helper from the service.
